requirements.txt
matplotlib==3.3.2
torch==1.7.0
numpy==1.19.4
torchvision==0.8.1
pandas==1.1.4
scipy==1.5.4
Pillow==8.0.1
